page-ranking module extracted
