+++ server/src/modules/page-ranking/db/entity/index.ts
@@ -1,0 +1,1 @@
+export type { PageEntity } from "./PageEntity";
