+++ server/src/modules/page-ranking/models/index.ts
@@ -1,0 +1,1 @@
+export type { Page } from "./Page";
