+++ server/src/modules/pages/Schemas.ts
@@ -17,16 +17,6 @@ export const pagesTypeDefs = gql`
     title: String
   }
 
-  type Page {
-    id: ID!
-    websiteId: String!
-    pageId: String!
-    commentsCount: String!
-    url: String!
-    meta: PageMeta
-    votes: PageVotes!
-  }
-
   type PageDetails {
     url: String!
     pageId: String!
@@ -36,8 +26,6 @@ export const pagesTypeDefs = gql`
   }
 
   extend type Query {
-    topCommentedPages(minCommentsVoteSum: Int!, limit: Int!): [Page!]
-    topRatedPages(minVoteSum: Int!, limit: Int!): [Page!]
     pageDetails(url: String!, fetchMetaIfNoCache: Boolean!): PageDetails!
   }
 `;
