+++ server/src/modules/pages/db/index.ts
@@ -1,5 +1,4 @@
 export type {
-  PageEntity,
   PageConfigEntity,
   PageDetailsEntity,
   UrlPageIdEntity,
@@ -8,5 +7,4 @@ export type {
 export type { UrlMeta, PageDetailsData } from "./page";
 
 export { PageRepository } from "./page";
-export { PageRankingRepository } from "./page-ranking";
 export { PageConfigRepository } from "./page-config";
