+++ server/src/modules/pages/models/index.ts
@@ -1,3 +1,2 @@
-export type { Page } from "./Page";
 export type { PageDetails } from "./PageDetails";
 export type { UrlId } from "./UrlId";
